% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fns.R
\name{test_timestwo}
\alias{test_timestwo}
\title{Times 2}
\usage{
test_timestwo(x)
}
\arguments{
\item{x}{A numerical vector}
}
\value{
The vector times 2
}
\description{
This function multiplies a vector by 2
}
